+++ TP - SBGames/Assets/Palettes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: afb33dc8941ff304190448850a91313f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
