+++ TP - SBGames/Assets/Palettes/New Palette 1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8c4821ebf46e5394daabfe1e4e61fad0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
