+++ TP - SBGames/Assets/Palettes/New Palette.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 461dfcc90e24c6043aff07d609a20b95
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
